+++ src/main/java/jscover/instrument/NodeProcessor.java
@@ -345,15 +345,16 @@ package jscover.instrument;
 import org.mozilla.javascript.Token;
 import org.mozilla.javascript.ast.*;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.SortedSet;
-import java.util.TreeSet;
+import java.util.*;
+
+import static jscover.instrument.IgnoreComment.IGNORE_END;
+import static jscover.instrument.IgnoreComment.IGNORE_START;
 
 //Function Coverage added by Howard Abrams, CA Technologies (HA-CA) - May 20 2013, tntim96
 class NodeProcessor {
     private StatementBuilder statementBuilder = new StatementBuilder();
     private SortedSet<Integer> validLines = new TreeSet<Integer>();
+    private LinkedList<IgnoreComment> ignores = new LinkedList<IgnoreComment>();
     private int functionNumber;// Function Coverage (HA-CA)
     private String fileName;
     private boolean includeFunctionCoverage;
@@ -373,6 +374,16 @@ class NodeProcessor {
     }
 
     boolean processNode(AstNode node) {
+        if (node instanceof Comment) {
+            String comment = ((Comment)node).getValue();
+            if (comment.startsWith(IGNORE_START)) {
+                ignores.add(new IgnoreComment(comment.substring(IGNORE_START.length() + 1), node.getLineno()));
+            } else if (comment.startsWith(IGNORE_END)) {
+                ignores.getLast().setEnd(node.getLineno());
+            }
+            return true;
+        }
+
         // Function Coverage (HA-CA), tntim96
         if (includeFunctionCoverage && node instanceof FunctionNode) {
             AstNode block = ((FunctionNode) node).getBody();
@@ -518,5 +529,9 @@ class NodeProcessor {
 
     public int getNumFunctions() {
     	return functionNumber;
+    }
+
+    public List<IgnoreComment> getIgnores() {
+        return ignores;
     }
 }