+++ src/main/java/jscover/instrument/ParseTreeInstrumenter.java
@@ -345,6 +345,7 @@ package jscover.instrument;
 import org.mozilla.javascript.ast.AstNode;
 import org.mozilla.javascript.ast.NodeVisitor;
 
+import java.util.List;
 import java.util.SortedSet;
 import java.util.logging.Logger;
 
@@ -369,6 +370,10 @@ class ParseTreeInstrumenter implements NodeVisitor {
 	// Function Coverage (HA-CA)
     public int getNumFunctions() {
     	return nodeProcessor.getNumFunctions();
+    }
+
+    public List<IgnoreComment> getIgnores() {
+        return nodeProcessor.getIgnores();
     }
 
     public boolean visit(AstNode node) {