+++ src/main/resources/jscoverage.js
@@ -524,8 +524,8 @@ function jscoverage_recalculateSummaryTab(cc) {
     var length = fileCC.length;
     var currentConditionalEnd = 0;
     var conditionals = null;
-    if (fileCC.conditionals) {
-      conditionals = fileCC.conditionals;
+    if (cc[file].conditionals) {
+      conditionals = cc[file].conditionals;
     }
     for (lineNumber = 0; lineNumber < length; lineNumber++) {
       var n = fileCC[lineNumber];
@@ -899,6 +899,7 @@ function jscoverage_checkbox_click() {
 function jscoverage_makeTable(lines) {
   var coverage = _$jscoverage[jscoverage_currentFile].lineData;
   var branchData = _$jscoverage[jscoverage_currentFile].branchData;
+  var conditionals = _$jscoverage[jscoverage_currentFile].conditionals;
 
   // this can happen if there is an error in the original JavaScript file
   if (! lines) {
@@ -933,9 +934,8 @@ function jscoverage_makeTable(lines) {
 
     if (lineNumber === currentConditionalEnd) {
       currentConditionalEnd = 0;
-    }
-    else if (currentConditionalEnd === 0 && coverage.conditionals && coverage.conditionals[lineNumber]) {
-      currentConditionalEnd = coverage.conditionals[lineNumber];
+    } else if (currentConditionalEnd === 0 && conditionals && conditionals[lineNumber]) {
+      currentConditionalEnd = conditionals[lineNumber];
     }
 
     var row = '<tr>';
@@ -944,17 +944,14 @@ function jscoverage_makeTable(lines) {
     if (timesExecuted !== undefined && timesExecuted !== null) {
       if (currentConditionalEnd !== 0) {
         row += '<td class="y numeric">';
-      }
-      else if (timesExecuted === 0) {
+      } else if (timesExecuted === 0) {
         row += '<td class="r numeric" id="line-' + lineNumber + '">';
-      }
-      else {
+      } else {
         row += '<td class="g numeric">';
       }
       row += timesExecuted;
       row += '</td>';
-    }
-    else {
+    } else {
       row += '<td></td>';
     }
 
@@ -963,14 +960,17 @@ function jscoverage_makeTable(lines) {
         var branchClass = '';
         var branchText = '&#160;';
         if (branchData[lineNumber] !== undefined && branchData[lineNumber] !== null) {
+          if (currentConditionalEnd !== 0) {
+            branchClass = 'y';
+          } else {
             branchClass = 'g';
             for (var conditionIndex = 0; conditionIndex < branchData[lineNumber].length; conditionIndex++) {
-                if (branchData[lineNumber][conditionIndex] !== undefined && branchData[lineNumber][conditionIndex] !== null && !branchData[lineNumber][conditionIndex].covered()) {
-                    branchClass = 'r';
-                    break;
-                }
+              if (branchData[lineNumber][conditionIndex] !== undefined && branchData[lineNumber][conditionIndex] !== null && !branchData[lineNumber][conditionIndex].covered()) {
+                branchClass = 'r';
+                break;
+              }
             }
-
+          }
         }
         if (branchClass === 'r') {
             branchText = '<a href="#" onclick="alert(buildBranchMessage(_$jscoverage[\''+jscoverage_currentFile+'\'].branchData[\''+lineNumber+'\']));">info</a>';
@@ -1090,7 +1090,7 @@ function jscoverage_recalculateSourceTab() {
             for (var i = 0; i < lines.length; i++)
                   lines[i] = jscoverage_html_escape(lines[i]);
               jscoverage_makeTable(lines);
-          }
+          };
           setTimeout(displaySource, 0);
           summaryThrobber.style.visibility = 'hidden';
         }
@@ -1160,8 +1160,7 @@ function jscoverage_selectTab(tab) {
     if (node.className !== 'disabled') {
       if (tabNum === tab) {
         node.className = 'selected';
-      }
-      else {
+      } else {
         node.className = '';
       }
     }
@@ -1178,8 +1177,7 @@ function jscoverage_selectTab(tab) {
 
     if (tabNum === tab) {
       node.className = 'selected TabPage';
-    }
-    else {
+    } else {
       node.className = 'TabPage';
     }
     tabNum++;
@@ -1222,8 +1220,7 @@ function jscoverage_tab_click(e) {
 //#JSCOVERAGE_IF
   if (e) {
     target = e.target;
-  }
-  else if (window.event) {
+  } else if (window.event) {
     // IE
     target = window.event.srcElement;
   }
@@ -1237,19 +1234,16 @@ function jscoverage_tab_click(e) {
       while (tbody.hasChildNodes()) {
         tbody.removeChild(tbody.firstChild);
       }
-    }
-    else if (target.id === 'sourceTab') {
+    } else if (target.id === 'sourceTab') {
       var sourceDiv = document.getElementById('sourceDiv');
       sourceDiv.innerHTML = '';
     }
     jscoverage_selectTab(target);
     if (target.id === 'summaryTab') {
       jscoverage_recalculateSummaryTab();
-    }
-    else if (target.id === 'sourceTab') {
+    } else if (target.id === 'sourceTab') {
       jscoverage_recalculateSourceTab();
-    }
-    else {
+    } else {
       jscoverage_endLengthyOperation();
     }
   }, 50);
@@ -1312,12 +1306,10 @@ function jscoverage_storeButton_click() {
           throw request.status;
         }
         message = request.responseText;
-      }
-      catch (e) {
+      } catch (e) {
         if (e.toString().search(/^\d{3}$/) === 0) {
           message = e + ': ' + request.responseText;
-        }
-        else {
+        } else {
           message = 'Could not connect to server: ' + e;
         }
       }
@@ -1354,12 +1346,10 @@ function jscoverage_stopButton_click() {
           throw request.status;
         }
         message = request.responseText;
-      }
-      catch (e) {
+      } catch (e) {
         if (e.toString().search(/^\d{3}$/) === 0) {
           message = e + ': ' + request.responseText;
-        }
-        else {
+        } else {
           message = 'Could not connect to server: ' + e;
         }
       }