+++ src/test-integration/resources/data/javascript-ignore/ignore-simple.js
@@ -1,0 +1,15 @@
+function createRequest() {
+//#JSCOVERAGE_IF window.XMLHttpRequest
+  if (window.XMLHttpRequest) {
+    return new XMLHttpRequest();
+  }
+//#JSCOVERAGE_ENDIF
+//#JSCOVERAGE_IF window.ActiveXObject
+  if (window.ActiveXObject) {
+    return new ActiveXObject('Msxml2.XMLHTTP');
+  }
+//#JSCOVERAGE_ENDIF
+  throw 'no XMLHttpRequest implementation available';
+}
+
+var request = createRequest();